requirements.txt
# Django framework
Django==5.2.8

# Environment variable management
django-environ==0.12.0

# Static file serving
whitenoise==6.9.0

# Data import/export functionality
django-import-export==4.3.2
openpyxl==3.1.5  # Excel file support

# Development tools
django-debug-toolbar==5.0.0
ruff==0.14.4
debugpy==1.8.11  # VSCode debugging support

# Django dependencies (explicitly listed)
asgiref==3.10.0
sqlparse==0.5.3

# PostgreSQL adapter
psycopg2-binary==2.9.10

# Google Gemini API
google-generativeai==0.8.3

# Production server
gunicorn==23.0.0
